+++ src/llik2.h
@@ -51,6 +51,5 @@ typedef struct stanLl {
   Eigen::Matrix<double, -1, -1> J;
 } stanLl;
 
-#define _smallIsOne(x) ((x) < sqrt(DBL_EPSILON) ? 1 : (x))
 #define _smallIsNotZero(x) (fabs(x) < sqrt(DBL_EPSILON) ? sqrt(DBL_EPSILON) : (x))
 #define _parIsProb(x) ((x) < sqrt(DBL_EPSILON) ? sqrt(DBL_EPSILON): ((x) > 0.999999999999999 ? 0.999999999999999 :(x)))